Factored OutputFile and SupportedLanguage into their own modules
